Test and implement Delete action method in BusStopController.

+++ GlogowskiBus.BLL/Concrete/BusStopService.cs
@@ -84,7 +84,10 @@ namespace GlogowskiBus.BLL.Concrete
 
         public void Delete(int id)
         {
+            if (unitOfWork.BusStopRepository.GetById(id) == null)
+                throw new Exception("Bus stop does not exist!");
             unitOfWork.BusStopRepository.Delete(id);
+            unitOfWork.Save();
         }
     }
 }
